+++ repository/user_repository.go
@@ -1,4 +1,4 @@
-package user
+package repository
 
 import (
 	"context"
